enemy movement modified

+++ code/Projectile.cpp
@@ -8,7 +8,7 @@ Projectile::Projectile(LPARAM lParam, float x2, float y2)
     direction[1] = HIWORD(lParam) - y2;
 
     // normalize
-    float divisor = sqrt(pow(direction[0], 2) + pow(direction[1], 2));
+    float divisor = (float)sqrt(pow(direction[0], 2) + pow(direction[1], 2));
     direction[0] /= divisor;
     direction[1] /= divisor;
     x = x2;
